refactor : Rename validator class and package to parser
