+++ src/main/kotlin/racingcar/view/input/NumberAndStepsInputView.kt
@@ -1,7 +1,7 @@
 package racingcar.view.input
 
-import racingcar.view.input.validator.InputValidator
-import racingcar.view.input.validator.IntInputValidator
+import racingcar.view.input.parser.InputParser
+import racingcar.view.input.parser.IntInputParser
 
 data class NumberAndSteps(val numberOfCar: Int, val stepsToTry: Int)
 
@@ -15,21 +15,21 @@ class NumberAndStepsInputView : InputView<NumberAndSteps> {
     }
 
     private fun readNumberOfCars() =
-        readFromConsole(MESSAGE_FOR_INPUT_NUMBER_OF_CARS, IntInputValidator(greaterThan = 0))
+        readFromConsole(MESSAGE_FOR_INPUT_NUMBER_OF_CARS, IntInputParser(greaterThan = 0))
 
-    private fun readStepsToTry() = readFromConsole(MESSAGE_FOR_INPUT_STEPS_TO_TRY, IntInputValidator(greaterThan = 0))
+    private fun readStepsToTry() = readFromConsole(MESSAGE_FOR_INPUT_STEPS_TO_TRY, IntInputParser(greaterThan = 0))
 
     companion object {
         private const val MESSAGE_FOR_INPUT_NUMBER_OF_CARS = "자동차 대수는 몇 대인가요?"
         private const val MESSAGE_FOR_INPUT_STEPS_TO_TRY = "시도할 횟수는 몇 회인가요?"
 
-        private fun <T> readFromConsole(message: String, validator: InputValidator<T>): T {
+        private fun <T> readFromConsole(message: String, inputParser: InputParser<T>): T {
             while (true) {
                 println(message)
                 try {
                     val oneLineInput = readLine()
                     if (!oneLineInput.isNullOrBlank()) {
-                        return validator.parseValue(oneLineInput)
+                        return inputParser.parseValue(oneLineInput)
                     }
                 } catch (e: Exception) {
                     println(e.message)
