+++ src/main/kotlin/racingcar/view/input/parser/InputParser.kt
@@ -1,0 +1,5 @@
+package racingcar.view.input.parser
+
+interface InputParser<T> {
+    fun parseValue(inputString: String?): T
+}
