--- src/main/kotlin/racingcar/view/input/validator/InputValidator.kt
@@ -1,5 +1,0 @@
-package racingcar.view.input.validator
-
-interface InputValidator<T> {
-    fun parseValue(inputString: String?): T
-}
